+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Set-SsisFolder and Remove-SsisFolder commands.
 - New-SsisFolder command.
 - Get-SsisFolder command.
 - Set-SsisCatalog command.